pyproject.toml
[project]
name = "detectops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "azure-storage-blob>=12.27.1",
    "bcrypt==4.0.1",
    "fastapi>=0.122.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
